refactored all span processor classes
