+++ integration/agent/api/src/main/java/com/splunk/rum/integration/agent/api/internal/SplunkRumAgentCore.kt
@@ -20,19 +20,18 @@ import android.app.Application
 import com.cisco.android.common.logger.Logger
 import com.cisco.android.common.logger.consumers.AndroidLogConsumer
 import com.splunk.rum.integration.agent.api.AgentConfiguration
-import com.splunk.rum.integration.agent.api.attributes.ErrorIdentifierAttributesSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.ErrorIdentifierAttributesSpanProcessor
 import com.splunk.rum.integration.agent.api.configuration.ConfigurationManager
 import com.splunk.rum.integration.agent.api.exporter.LoggerSpanExporter
 import com.splunk.rum.integration.agent.api.extension.toResource
-import com.splunk.rum.integration.agent.api.internal.processors.GlobalAttributeSpanProcessor
-import com.splunk.rum.integration.agent.api.sessionId.SessionIdSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.GlobalAttributeSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.SessionIdSpanProcessor
 import com.splunk.rum.integration.agent.api.sessionId.SessionStartEventManager
-import com.splunk.rum.integration.agent.api.user.UserIdSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.UserIdSpanProcessor
 import com.splunk.rum.integration.agent.internal.AgentIntegration
-import com.splunk.rum.integration.agent.internal.span.AppStartSpanProcessor
-import com.splunk.rum.integration.agent.internal.span.SplunkInternalGlobalAttributeSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.AppStartSpanProcessor
+import com.splunk.rum.integration.agent.internal.processor.SplunkInternalGlobalAttributeSpanProcessor
 import com.splunk.rum.integration.agent.internal.user.IUserManager
-import com.splunk.rum.integration.agent.internal.user.UserManager
 import com.splunk.rum.integration.agent.module.ModuleConfiguration
 import com.splunk.sdk.common.otel.OpenTelemetryInitializer
 import com.splunk.sdk.common.storage.AgentStorage
