+++ integration/agent/internal/build.gradle.kts
@@ -38,6 +38,7 @@ dependencies {
     implementation(Dependencies.SessionReplay.commonHttp)
     implementation(Dependencies.SessionReplay.commonStorage)
     implementation(Dependencies.SessionReplay.commonUtils)
+    implementation(project(":common:utils"))
 
     compileOnly(Dependencies.Android.Compose.ui)
 }
