+++ integration/agent/internal/src/main/java/com/splunk/rum/integration/agent/internal/processor/UserIdSpanProcessor.kt
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 
-package com.splunk.rum.integration.agent.api.user
+package com.splunk.rum.integration.agent.internal.processor
 
-import com.splunk.rum.integration.agent.api.attributes.AttributeConstants.USER_ID_KEY
+import com.splunk.rum.integration.agent.internal.attributes.AttributeConstants.USER_ID_KEY
 import com.splunk.rum.integration.agent.internal.user.IUserManager
 import io.opentelemetry.context.Context
 import io.opentelemetry.sdk.trace.ReadWriteSpan
 import io.opentelemetry.sdk.trace.ReadableSpan
 import io.opentelemetry.sdk.trace.SpanProcessor
 
-internal class UserIdSpanProcessor(private val userManager: IUserManager) : SpanProcessor {
+class UserIdSpanProcessor(private val userManager: IUserManager) : SpanProcessor {
     override fun onStart(parentContext: Context, span: ReadWriteSpan) {
         val userId = userManager.userId ?: return
         span.setAttribute(USER_ID_KEY, userId)
